requirements.txt

# FastAPI / Core
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# PDF Processing
PyMuPDF==1.23.21

# LangChain Core
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.3
langchain-huggingface>=0.0.1

# Embeddings & Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0
huggingface-hub>=0.23.0

# Utilities
tqdm>=4.66.1
